+++ src/library/Hooks/useOutsideAlerter/index.tsx
@@ -1,13 +1,10 @@
 // Copyright 2023 @paritytech/polkadot-staking-dashboard authors & contributors
 // SPDX-License-Identifier: GPL-3.0-only
 
-import type { FC, RefObject } from 'react';
-import { useEffect } from 'react';
+import { useEffect, type RefObject } from 'react';
 import type { AnyFunction, AnyJson } from 'types';
 
-/*
- * A hook that alerts clicks outside of the passed ref.
- */
+// A hook that alerts clicks outside of the passed ref.
 export const useOutsideAlerter = (
   ref: RefObject<HTMLElement>,
   callback: AnyFunction,
@@ -30,21 +27,3 @@ export const useOutsideAlerter = (
     };
   }, [ref]);
 };
-
-// A pure function that applies an arbitrary amount of context providers to a wrapped
-// component.
-export const withProviders = (
-  providers: (FC<AnyJson> | [FC<AnyJson>, AnyJson])[],
-  Wrapped: FC
-) =>
-  providers.reduceRight(
-    (acc, prov) => {
-      if (Array.isArray(prov)) {
-        const Provider = prov[0];
-        return <Provider {...prov[1]}>{acc}</Provider>;
-      }
-      const Provider = prov;
-      return <Provider>{acc}</Provider>;
-    },
-    <Wrapped />
-  );
