+++ src/library/Menu/index.tsx
@@ -3,9 +3,9 @@
 
 import { useEffect, useRef } from 'react';
 import { useMenu } from 'contexts/Menu';
-import { useOutsideAlerter } from 'library/Hooks';
 import { ItemWrapper, Wrapper } from './Wrappers';
 import type { AnyJson } from 'types';
+import { useOutsideAlerter } from 'library/Hooks/useOutsideAlerter';
 
 export const Menu = () => {
   const menu = useMenu();
